+++ lib/marquee_direction.dart
@@ -1,4 +1,13 @@
 enum MarqueeDirection {
+  /// from right to left
   rtl,
+
+  /// from left to right
   ltr,
+
+  /// from bottom to top
+  btt,
+
+  /// from top to bottom
+  ttb,
 }